requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
lxml==4.8.0
numpy==1.22.3
olefile==0.46
Pillow==9.1.0
requests==2.27.1
reverse-geocoder==1.5.1
scipy==1.8.0
urllib3==1.26.9
